For our last trick, Google for a library called SweetAlert2. Very simply, this library give us sweet alert boxes, like this. And you can customize it in a lot of ways, like having a 'Yes' and 'Cancel' button.
We're going to use SweetAlert so that when we click the delete icon, an alert opens so the user can confirm the delete before we actually do it.

SweetAlert: Basic Usage

To get this installed, go to the CDN. Copy the JavaScript file first. This time, instead of putting this in our base layout, we'll add the JavaScript to just this page: index.html.twig. Add the <script src='https://cdn.jsdelivr.net/sweetalert2/6.1.0/sweetalert2.min.js'>

This also comes with a CSS file: copy that too. Back in index.html.twig, override a block called stylesheets and add the endblock. Call parent() to include the normal stylesheets, and then add the link tag with this path.

Perfect!

This library exposes a global swal() function. Copy the timer example - it's somewhat similar to what we want. Then, open RepLogApp.js. Remember, whenever we reference a global object, we like to pass it into our self-executing function. You don't need to do this, but it's super hipster. Pass swal at the bottom and also swal on top.

If you want some auto-completion on that library, you can of course select it and hit option+enter or alt+enter to tell PhpStorm to download it.