Down in handleRepLogDelete, here's the plan. First, we'll open the alert. And then, when the user clicks 'OK', we'll run all of the code below that actually deletes the RepLog. To prepare for that, isolate all of that into its own new method: _deleteRepLog with a {dollar_sign}link argument.
This doesn't change anything: we could still just call this function directly from above. But instead, paste the SweetAlert code and update the title - 'Delete this log' - and the text - 'Did you not actually lift this?'. And remove the timer option. Instead, add showCancelButton: true.

With just that, we should be able to refresh, and... oh! Error!

    swal is not defined