When we call swal(), guess what it returns? A promise! A freaking Promise! We can tell because the code has a .then chained to it, with two arguments. The first argument is the function that's called on success, and the second is called when the Promise is rejected. But, we already knew that.
Specifically, for SweetAlert, the success, or resolved handler is called if we click 'OK', and the reject handler is called if we click 'Cancel'. Easy! Above the swal() call, add var self = this. Then, inside the success handler, use self._deleteRepLog({dollar_sign}link)

Down in the reject function, we don't need to do anything. Just call console.log('canceled').

Let's try it! Refresh, click the trash icon and hit Cancel. Yea, there's the log! Now hit 'OK'. It deletes it! Guys, this is why understanding promises is so important.