More realistically, sometimes - instead of doing my work after the alert closes, I like to do my work, my AJAX call, inside of preConfirm. After all, SweetAlert shows the user a pretty fancy loading icon while they're waiting. Let's do that here - it's super easy!
Move the self._deleteRepLog() call up into the preConfirm function and return it. Then get rid of the .then() entirely.

This is totally legal, as long as the _deleteRepLog() function returns a Promise. In other words, as long as we return {dollar_sign}.ajax(), SweetAlert will be happy.

We can still keep the catch here, because if you hit cancel, that will still reject the promise and call .catch(). Head back, refresh, and click delete. You should see the loading icon for just a moment, while our AJAX call finishes. Hit 'Ok'! Beautiful!